requirements.txt
# Main training requirements
tensorflow-gpu == 1.14.0
numpy == 1.15.4
progressbar2
pandas
six
pyxdg
attrdict
absl-py

# Requirements for building native_client files
setuptools

# Requirements for transcribe.py
webrtcvad

# Requirements for importers
sox
bs4
requests
librosa
soundfile

# Miscellaneous scripts
paramiko >= 2.1
scipy
matplotlib

# My
pydub
wget
